Major edits to grouping and aggregation logic
Make major changes to how grouping and aggregation is handled based on
whether or not the field of an event has a grouping context and
instance.

Additionally, add GroupAggregateFields to act as a central place to
collect the configured group and aggregate fields supplied by the user.

+++ warehouse/query-core/src/test/java/datawave/query/common/grouping/CountAggregatorTest.java
@@ -25,7 +25,7 @@ public class CountAggregatorTest {
      */
     @Test
     public void testInitialCount() {
-        assertCount(0);
+        assertCount(0L);
     }
     
     /**
@@ -34,17 +34,17 @@ public class CountAggregatorTest {
     @Test
     public void testIncrementingCount() {
         aggregator.aggregate(value);
-        assertCount(1);
+        assertCount(1L);
         
         aggregator.aggregate(value);
-        assertCount(2);
+        assertCount(2L);
         
         aggregator.aggregate(value);
         aggregator.aggregate(value);
-        assertCount(4);
+        assertCount(4L);
     }
     
-    private void assertCount(Integer count) {
+    private void assertCount(Long count) {
         assertEquals(count, aggregator.getAggregation());
     }
